mandatory: checking if int, INT_MAX & INT_MIN, freeing all memory for now

+++ libft/libft.h
@@ -71,14 +71,14 @@ t_list	*ft_lstnew(void *content);
 int		ft_lstsize(t_list *lst);
 
 //Other functions
-char	*ft_hex_itoa(unsigned int n);
-char	*ft_pointer_itoa(unsigned long int n);
-char	*ft_uitoa(unsigned int n);
-size_t	ft_strlen_char(const char *str, char a);
-void	ft_printf_array(char **array);
-void	ft_free_array(char **array);
-char	*ft_strtrim_mod(char *s1, char const *set);
-void	ft_free_int(int **array);
-int		ft_atoi_mod(char *nptr);
+char		*ft_hex_itoa(unsigned int n);
+char		*ft_pointer_itoa(unsigned long int n);
+char		*ft_uitoa(unsigned int n);
+size_t		ft_strlen_char(const char *str, char a);
+void		ft_printf_array(char **array);
+void		ft_free_array(char **array);
+char		*ft_strtrim_mod(char *s1, char const *set);
+void		ft_free_int(int **array);
+long int	ft_atoi_mod(char *nptr);
 
 #endif
